+++ bzst-dip-java-client/src/main/java/software/xdev/bzst/dip/client/model/configuration/BzstDipConfigurationBuilder.java
@@ -52,6 +52,10 @@ public class BzstDipConfigurationBuilder
 	 */
 	private String taxID;
 	/**
+	 * @see BzstDipConfiguration#isValidateTaxID()
+	 */
+	private Boolean validateTaxID;
+	/**
 	 * @see BzstDipConfiguration#getTaxNumber()
 	 */
 	private String taxNumber;
@@ -159,6 +163,16 @@ public class BzstDipConfigurationBuilder
 	public BzstDipConfigurationBuilder setTaxID(final String taxID)
 	{
 		this.taxID = taxID;
+		return this;
+	}
+	
+	/**
+	 * @param validateTaxID {@link #validateTaxID}
+	 * @return itself
+	 */
+	public BzstDipConfigurationBuilder setValidateTaxID(final Boolean validateTaxID)
+	{
+		this.validateTaxID = validateTaxID;
 		return this;
 	}
 	
@@ -330,6 +344,10 @@ public class BzstDipConfigurationBuilder
 				"certificate"),
 			this.getSetPropertyOrReadFromFile(this.clientId, PropertiesSupplier.PROPERTY_NAME_CLIENT_ID),
 			this.getSetPropertyOrReadFromFile(this.taxID, PropertiesSupplier.PROPERTY_NAME_TAX_ID),
+			this.getSetPropertyOrReadFromFileBoolean(
+				this.validateTaxID,
+				PropertiesSupplier.PROPERTY_NAME_TAX_ID_VALIDATE,
+				true),
 			this.getSetPropertyOrReadFromFile(this.taxNumber, PropertiesSupplier.PROPERTY_NAME_TAX_NUMBER),
 			this.getSetPropertyOrReadFromFile(
 				this.realmEnvironmentBaseUrl,
@@ -463,6 +481,21 @@ public class BzstDipConfigurationBuilder
 			String.valueOf(defaultValue)));
 	}
 	
+	private Boolean getSetPropertyOrReadFromFileBoolean(
+		final Boolean builderProperty,
+		final String propertyNameInFile,
+		final Boolean defaultValue)
+	{
+		if(builderProperty != null)
+		{
+			return builderProperty;
+		}
+		return Boolean.parseBoolean(this.getSetPropertyOrReadFromFile(
+			null,
+			propertyNameInFile,
+			String.valueOf(defaultValue)));
+	}
+	
 	private Duration getSetPropertyOrReadFromFileDuration(
 		final Duration builderProperty,
 		final String propertyNameInFile,
@@ -538,7 +571,6 @@ public class BzstDipConfigurationBuilder
 		final String propertyNameInFile,
 		final String defaultValue)
 	{
-		
 		if(builderProperty != null)
 		{
 			return builderProperty;